+++ src/OneOffPayments/TabBuilder.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Laravel\Cashier\OneOffPayments;
+
+class TabBuilder
+{
+}
