+++ src/OneOffPayments/TabItemBuilder.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Laravel\Cashier\OneOffPayments;
+
+class TabItemBuilder
+{
+}
