+++ src/OneOffPayments/TabItemCollection.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Laravel\Cashier\OneOffPayments;
+
+use Illuminate\Database\Eloquent\Collection;
+
+class TabItemCollection extends Collection
+{
+}
